refactored the functions that add edit and delete workouts to be in the user.ts file to make working with them easier
